Update Inputs in InfoInputPage to components
