+++ components/ClubCreationPage/index.tsx
@@ -2,7 +2,6 @@ import * as S from './style'
 import { useSelector } from 'react-redux'
 import { RootState } from '@/store'
 import ClubKindSelection from './Steps/ClubKindSelection'
-import { resetPage } from '@/store/clubCreationPage'
 import ClubInfoInput from './Steps/ClubInfoInput'
 import Confirm from './Steps/Confirm'
 import { useRouter } from 'next/router'
